requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==4.11.3
mysqlclient==2.1.1
pytz==2022.6
sqlparse==0.4.3
typed-ast==1.5.2
typing_extensions==4.4.0
zipp==3.7.0
